+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 
 ## [Unreleased]
 
+### Added 
+- Token Exchange : Allow requested_token_type=urn:ietf:params:oauth:token-type:refresh_token with offline access token and created access token containing offline_access
+
 ### Fixed
 - Fix bugs in Patch and query of members (Group SCIM)
 - Support OID-Fed Automatic Client registration for client_credentials grant (Keycloak as OP)