+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
-## [Unreleased]
+## [26.6.6-1.2] -2026-09-15
 
 ### Added 
 - Token Exchange : Allow requested_token_type=urn:ietf:params:oauth:token-type:refresh_token with offline access token and created access token containing offline_access